+++ example/lib/views/events_planner_multi_columns_view.dart
@@ -1,5 +1,4 @@
 import 'package:example/data.dart';
-import 'package:example/extension.dart';
 import 'package:flutter/material.dart';
 import 'package:infinite_calendar_view/infinite_calendar_view.dart';
 import 'package:intl/intl.dart';
@@ -24,36 +23,54 @@ class EventsPlannerMultiColumnView extends StatelessWidget {
 
     return EventsPlanner(
       controller: controller,
-      daysShowed: 1,
+      daysShowed: 2,
       heightPerMinute: heightPerMinute,
       initialVerticalScrollOffset: initialVerticalScrollOffset,
       daySeparationWidth: 10,
       columnsParam: ColumnsParam(
         columns: 3,
-        columnsLabels: ["Tennis", "Foot", "Bad"],
         columnsWidthRatio: [1 / 3, 1 / 3, 1 / 3],
-        columnsColors: [
-          Colors.yellow.pastel,
-          Colors.green.pastel,
-          Colors.blueAccent.pastel,
-        ],
+        columnHeaderBuilder: (day, isToday, columIndex, columnWidth) {
+          return DefaultColumnHeader(
+            backgroundColor: Colors.transparent,
+            foregroundColor: Colors.white,
+            columnText: ["Tennis", "Foot", "Bad"].elementAt(columIndex),
+            columnWidth: columnWidth,
+          );
+        },
       ),
-      fullDayParam: FullDayParam(
-        fullDayEventsBarVisibility: false,
-      ),
+      fullDayParam: FullDayParam(fullDayEventsBarVisibility: false),
       daysHeaderParam: DaysHeaderParam(
         daysHeaderHeight: 80,
-        daysHeaderVisibility: true,
-        daysHeaderColor: Theme.of(context).appBarTheme.backgroundColor,
         dayHeaderBuilder: (day, isToday) {
           return DefaultDayHeader(
             dayText: DateFormat("E d").format(day),
             isToday: isToday,
-            foregroundColor:
-                isDarkMode ? Theme.of(context).colorScheme.primary : null,
+          );
+        },
+      ),
+      dayParam: DayParam(
+        dayEventBuilder: (event, height, width, heightPerMinute) {
+          return DefaultDayEvent(
+            height: height,
+            width: width,
+            color: event.color,
+            textColor: event.textColor,
+            title: event.title,
+            description: getSlotHourText(event.startTime, event.endTime!),
           );
         },
       ),
     );
+  }
+
+  String getSlotHourText(DateTime start, DateTime end) {
+    return start.hour.toString().padLeft(2, '0') +
+        ":" +
+        start.hour.toString().padLeft(2, '0') +
+        "\n" +
+        end.hour.toString().padLeft(2, '0') +
+        ":" +
+        end.hour.toString().padLeft(2, '0');
   }
 }